🛠️ Refactoring TV scrapping flow
♻️ Starting process of removing interfaces as dependencies in core processes
